game: thinktime should be an int, see: http://forums.warchest.com/showthread.php/12761-A-pesky-bug-fix
misc: typo fixes

+++ src/game/g_main.c
@@ -4222,7 +4222,7 @@ void CheckCvars(void)
  */
 void G_RunThink(gentity_t *ent)
 {
-	float thinktime;
+	int thinktime;
 
 	// If paused, push nextthink
 	if (level.match_pause != PAUSE_NONE && (ent - g_entities) >= g_maxclients.integer &&
